fix(seccomp): make BlockedFamily and OnBlockAction platform-neutral
BlockedFamily, ParseFamily, DefaultBlockedFamilies are pure config
helpers consumed by internal/config (platform-neutral). The original
//go:build linux && cgo tag broke cross-platform and CGO_ENABLED=0
builds for any consumer.

Extract OnBlockAction + ParseOnBlock to a new actions.go (no tag)
and drop the tag from families.go. filter.go stays linux/cgo for now;
extracting its pure parts is a separate concern.

Found by roborev (max reasoning) on commit 0f5b68e.
